+++ src/js/view/initialView.js
