--- src/js/view/winnerViewControl.js
@@ -1,5 +1,0 @@
-import { $ } from "./getElement.js";
-
-export const setWinnerText = winners => {
-  $("racing-winner").innerHTML = winners.join(", ");
-};
